worker/requirements.txt
redis
flask-restful
PyYAML
xlrd
requests
python-dateutil